requirements.txt
# Core dependencies
numpy>=1.20.0
pysam>=0.19.0
PyYAML>=6.0

# Optional read simulation tools (install separately)
# ART (Illumina read simulator) - recommended
# BBTools - alternative read simulator

# Required external tools (install separately)
# BWA-MEM or minimap2 - for read mapping
# samtools - for BAM processing